+++ src/main/java/com/via/agencia_viagens/model/PrecoNormalStrategy.java
@@ -1,8 +1,10 @@
 package com.via.agencia_viagens.model;
 
+import java.math.BigDecimal;
+
 public class PrecoNormalStrategy implements CalculoPrecoStrategy {
     @Override
-    public double calcularPreco(double precoBase) {
+    public BigDecimal calcularPreco(BigDecimal precoBase) {
         return precoBase;
     }
 }
